python3 syntax fix

+++ scripts/models.py
@@ -11,7 +11,7 @@ from glob import glob
 
 def Deconv(inputs, f_dim_in, dim, net, batch_size, f_dim_out = None, stride = 2 ):
 	if f_dim_out is None: 
-		f_dim_out = f_dim_in/2 
+		f_dim_out = int(f_dim_in/2)
 	return tl.layers.DeConv3dLayer(inputs,
 								shape = [4, 4, 4, f_dim_out, f_dim_in],
 								output_shape = [batch_size, dim, dim, dim, f_dim_out],
@@ -21,7 +21,7 @@ def Deconv(inputs, f_dim_in, dim, net, batch_size, f_dim_out = None, stride = 2 
 
 def Conv3D(inputs, f_dim_out, net, f_dim_in = None, batch_norm = False, is_train = True):
 	if f_dim_in is None: 
-		f_dim_in = f_dim_out/2
+		f_dim_in = int(f_dim_out/2)
 	layer = tl.layers.Conv3dLayer(inputs, 
 								   shape=[4, 4, 4, f_dim_in, f_dim_out],
 								   W_init = tf.random_normal_initializer(stddev=0.02),
